CPU: fix branch page crossing + offset calulation (#1)
Page crossing and offset during a branch are computed from the next instruction onward, because the 6502 advances the PC during fetching of the opcode and operand
This means, that a branch at CFFE to D0xx doesnt count as crossing a page boundary. Advancement of the PC also causes the base address of the branch to shift, as the offset is not applied to where the branch opcode is, but where the PC is after reading the whole branch instruction.
I also changed how the PC is set to avoid casting the PC to i16 and back to u16, just to get rid of a potential headache with the high bit of the PC being set and potentially causing the i16 representation to be negative (depending on how jakt/c++ do the cast in the background).
Since I didnt have jakt around when writing these changes, I'll mark the PR as a draft for now.
Real PR now since I had a chance to pass it through jakt and fix all the compiler errors.