Implement ADC, SBC, CMP with flags. (#7)
ADC is `a + b + c`
SBC is `a - b - ~c`

The `+ c` is implemented as the `carry_in` to the ripple carry adder and is cheap; ADC wires it to the carry flag. Since the 6502 uses twos-complement to implement negative numbers, `a - b = a + (~b + 1)`. When `carry` is `1`, `a - b - ~1` becomes `a + (~b + 1) + 0 = a + ~b + 1` and when `carry` is `0`, `a - b - ~0` becomes `a + (~b + 1) - 1 = a + ~b + 0`.  To implement subtraction, SBC inverts `b` and then performs ADC!

In general, `overflow` happens if the `carry_in` to the sum for bit 7 is different to the `carry_out` from bit 7. The implementation I used follows the C++ version described in http://www.righto.com/2012/12/the-6502-overflow-flag-explained.html