CPU: Add more wrapping support to adressing mode helpers
All addressing modes that are indexed, can exceed the range of u8/u16.
The 6502 deals with that by wrapping around, basically ignoring the overflown
bits. Jakt wants us to be safe and checks at runtime, that our calculations don't
exceed the range of the type. Since we know what we are doing, and C++ should do the
truncating for us when converting the type down after promotion, we use unchecked_add
to take advantage of that and dont get a runtime abort.