CPU: Make JMP be more accurate.
JMP is straight forward unless you have to cross page boundaries during
the indirect read, because the 6502 will latch the high byte and wrap
around on the current page instead of continuing on the next page.
While you should never read the indirect vector from the last byte of
the page to save your sanity, nothing stops you from doing that and
nestest has a test for this too!