CPU improvements and fixes (#12)
This PR includes several edits:
- ~~There was a typo in the big match{} in run_opcode~~ This was already
fixed in the `wip` branch
- #5 increased the PC offset in JSR by 1 (to pc + 3). While this fixed
the symptoms of RTS returning to an operand rather than an opcode, the
error was in RTS not incrementing the pc by 1 after pulling it from the
stack. Fixed now.
- Stack access was all over the code and warranted some helpers. Added
them. Also removed CPU::pull_pc() as it degenerated to `.pc =
pull_word()` and it was used only in RTS and RTI, with RTS needing to
increment the PC anyways.
- BRK was pushing the current PC on the stack, not PC + 2. This would
have resulted in a BRK loop, since RTI returns to the pushed PC.
- The addressing modes (and/or read_memory helpers) lack wrapping
support, which should be easy to implement:
    - [x] zero page, zero page x, zero page y
    - [x] indirect zero page x
    - [x] indirect zero page y
    - [x] absolute 
    - [x] absolute x, absolute y
- JMP is a fun one when the address for indirection starts on the last
byte of the page, as it does not continue the read on the next page, but
wraps around on the current page
- Added missing addressing mode helpers, and used them in places not
previously used.