PPU render_sprite_pixel: rudimentary optimization
Did some very simplistic optimization of this function. I'm not happy about
the unchecked math for sprite_attr but when profiling, these lines were
quite hot. There is certainly a better way to optimize this function
specifically, but there may also be optimizations possible in Jakt itself.
I'm just here as a hobbyist programmer though so it's going to take me a
lot more time to wrap my head around all of this.