Improve SDL rendering (#39)
By changing the way we pass the framebuffer information to SDL, we can
improve performance.

This change should open up a lot of potential for further render
improvement.

Some of Video.jakt is written in inline cpp, either because I couldnt
figure out how to use address-of in jakt, or couldn't make cpp happy
with the objects it got, or how to include constants like
SDL_TextureAccess_Streaming. This might also be some limitation in Jakt
for now.

Side note: For whatever reason, at least on my laptop, starting jakt_nes
via SSH makes the emulator gain a huge chunk of performance, to the
point where it might be too fast. I cannot explain that behaviour, so it
should not be relied upon.
But this shows that we might still be held back by (the way we call)
SDL2.