CPU - Opcode 1c
This is an unofficial opcode which performs NOP but consumes 4 cycles

This is referred to as "IGN a,X" in the nesdev unofficial opcodes, in that it
reads from memory at the specified address but then does nothing, with no
effect on register or flags.

IGN a,X can take 4 or 5 cycles, this implementation is 4 until we find a
reason to do 5.