CI: Update setup-jakt to version that does not use setup-cpp action
The artifacts from that action's recent releases do not like being
relocated the toolchain cache location, and resulted in stdlib headers
not being found properly from the underlying VM instance in Github
Actions. The new revision relies on the default compiler for GA VMs
supporting the subset of C++20 needed to compile jakt, which seems to
be the case for the tags we're currently using.